Fix device advertisment not sending correct service info
This is the really important part in order to get apps to detect the
device! Without these being advertised many apps won't bother connecting
in order to interrogate it. I suspect this was the problem where apps
previously wouldn't detect it properly. I tested on MyWhoosh and it is
detected.